backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-allauth>=0.57
PyJWT>=2.8
requests>=2.31
cryptography>=46.0
Pillow>=10.0.0

# WebSocket Support
channels>=4.0
daphne>=4.0

# ML and Resume Processing Dependencies
groq>=0.4.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
pandas>=2.0.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
